+++ stwo_cairo_prover/crates/cairo-air/src/cairo_components.rs
@@ -1,0 +1,308 @@
+use itertools::chain;
+use stwo::core::air::Component;
+use stwo::prover::backend::simd::SimdBackend;
+use stwo::prover::ComponentProver;
+use stwo_constraint_framework::preprocessed_columns::PreProcessedColumnId;
+use stwo_constraint_framework::TraceLocationAllocator;
+
+use crate::cairo_claim::CairoClaim;
+use crate::cairo_interaction_claim::CairoInteractionClaim;
+use crate::cairo_interaction_elements::CairoInteractionElements;
+use crate::components::{
+    indented_component_display, memory_address_to_id, memory_id_to_big, verify_bitwise_xor_4,
+    verify_bitwise_xor_7, verify_bitwise_xor_8, verify_bitwise_xor_8_b, verify_bitwise_xor_9,
+    verify_instruction,
+};
+use crate::opcodes_air::OpcodeComponents;
+use crate::builtins_air::BuiltinComponents;
+use crate::range_checks_air::RangeChecksComponents;
+use crate::blake::air::BlakeContextComponents;
+use crate::pedersen::air::PedersenContextComponents;
+use crate::poseidon::air::PoseidonContextComponents;
+pub struct CairoComponents {
+    pub opcodes: OpcodeComponents,
+    pub verify_instruction: verify_instruction::Component,
+    pub blake_context: BlakeContextComponents,
+    pub builtins: BuiltinComponents,
+    pub pedersen_context: PedersenContextComponents,
+    pub poseidon_context: PoseidonContextComponents,
+    pub memory_address_to_id: memory_address_to_id::Component,
+    pub memory_id_to_value: (
+        Vec<memory_id_to_big::BigComponent>,
+        memory_id_to_big::SmallComponent,
+    ),
+    pub range_checks: RangeChecksComponents,
+    pub verify_bitwise_xor_4: verify_bitwise_xor_4::Component,
+    pub verify_bitwise_xor_7: verify_bitwise_xor_7::Component,
+    pub verify_bitwise_xor_8: verify_bitwise_xor_8::Component,
+    pub verify_bitwise_xor_8_b: verify_bitwise_xor_8_b::Component,
+    pub verify_bitwise_xor_9: verify_bitwise_xor_9::Component,
+    // ...
+}
+impl CairoComponents {
+    pub fn new(
+        cairo_claim: &CairoClaim,
+        interaction_elements: &CairoInteractionElements,
+        interaction_claim: &CairoInteractionClaim,
+        // Describes the structure of the preprocessed trace. Sensitive to order.
+        preprocessed_column_ids: &[PreProcessedColumnId],
+    ) -> Self {
+        let tree_span_provider =
+            &mut TraceLocationAllocator::new_with_preprocessed_columns(preprocessed_column_ids);
+
+        let opcode_components = OpcodeComponents::new(
+            tree_span_provider,
+            &cairo_claim.opcodes,
+            interaction_elements,
+            &interaction_claim.opcodes,
+        );
+
+        let verify_instruction_component = verify_instruction::Component::new(
+            tree_span_provider,
+            verify_instruction::Eval {
+                claim: cairo_claim.verify_instruction,
+                memory_address_to_id_lookup_elements: interaction_elements
+                    .memory_address_to_id
+                    .clone(),
+                verify_instruction_lookup_elements: interaction_elements.verify_instruction.clone(),
+                memory_id_to_big_lookup_elements: interaction_elements.memory_id_to_value.clone(),
+                range_check_4_3_lookup_elements: interaction_elements.range_checks.rc_4_3.clone(),
+                range_check_7_2_5_lookup_elements: interaction_elements
+                    .range_checks
+                    .rc_7_2_5
+                    .clone(),
+            },
+            interaction_claim.verify_instruction.claimed_sum,
+        );
+
+        let blake_context = BlakeContextComponents::new(
+            tree_span_provider,
+            &cairo_claim.blake_context,
+            interaction_elements,
+            &interaction_claim.blake_context,
+        );
+        let builtin_components = BuiltinComponents::new(
+            tree_span_provider,
+            &cairo_claim.builtins,
+            interaction_elements,
+            &interaction_claim.builtins,
+        );
+        let pedersen_context = PedersenContextComponents::new(
+            tree_span_provider,
+            &cairo_claim.pedersen_context,
+            interaction_elements,
+            &interaction_claim.pedersen_context,
+        );
+        let poseidon_context = PoseidonContextComponents::new(
+            tree_span_provider,
+            &cairo_claim.poseidon_context,
+            interaction_elements,
+            &interaction_claim.poseidon_context,
+        );
+        let memory_address_to_id_component = memory_address_to_id::Component::new(
+            tree_span_provider,
+            memory_address_to_id::Eval::new(
+                cairo_claim.memory_address_to_id.clone(),
+                interaction_elements.memory_address_to_id.clone(),
+            ),
+            interaction_claim.memory_address_to_id.clone().claimed_sum,
+        );
+
+        let memory_id_to_value_components = memory_id_to_big::big_components_from_claim(
+            &cairo_claim.memory_id_to_value.big_log_sizes,
+            &interaction_claim.memory_id_to_value.big_claimed_sums,
+            &interaction_elements.memory_id_to_value,
+            &interaction_elements.range_checks.rc_9_9,
+            &interaction_elements.range_checks.rc_9_9_b,
+            &interaction_elements.range_checks.rc_9_9_c,
+            &interaction_elements.range_checks.rc_9_9_d,
+            &interaction_elements.range_checks.rc_9_9_e,
+            &interaction_elements.range_checks.rc_9_9_f,
+            &interaction_elements.range_checks.rc_9_9_g,
+            &interaction_elements.range_checks.rc_9_9_h,
+            tree_span_provider,
+        );
+        let small_memory_id_to_value_component = memory_id_to_big::SmallComponent::new(
+            tree_span_provider,
+            memory_id_to_big::SmallEval::new(
+                cairo_claim.memory_id_to_value.clone(),
+                interaction_elements.memory_id_to_value.clone(),
+                interaction_elements.range_checks.rc_9_9.clone(),
+                interaction_elements.range_checks.rc_9_9_b.clone(),
+                interaction_elements.range_checks.rc_9_9_c.clone(),
+                interaction_elements.range_checks.rc_9_9_d.clone(),
+            ),
+            interaction_claim
+                .memory_id_to_value
+                .clone()
+                .small_claimed_sum,
+        );
+        let range_checks_component = RangeChecksComponents::new(
+            tree_span_provider,
+            &interaction_elements.range_checks,
+            &interaction_claim.range_checks,
+        );
+        let verify_bitwise_xor_4_component = verify_bitwise_xor_4::Component::new(
+            tree_span_provider,
+            verify_bitwise_xor_4::Eval {
+                claim: cairo_claim.verify_bitwise_xor_4,
+                verify_bitwise_xor_4_lookup_elements: interaction_elements
+                    .verify_bitwise_xor_4
+                    .clone(),
+            },
+            interaction_claim.verify_bitwise_xor_4.claimed_sum,
+        );
+        let verify_bitwise_xor_7_component = verify_bitwise_xor_7::Component::new(
+            tree_span_provider,
+            verify_bitwise_xor_7::Eval {
+                claim: cairo_claim.verify_bitwise_xor_7,
+                verify_bitwise_xor_7_lookup_elements: interaction_elements
+                    .verify_bitwise_xor_7
+                    .clone(),
+            },
+            interaction_claim.verify_bitwise_xor_7.claimed_sum,
+        );
+        let verify_bitwise_xor_8_component = verify_bitwise_xor_8::Component::new(
+            tree_span_provider,
+            verify_bitwise_xor_8::Eval {
+                claim: cairo_claim.verify_bitwise_xor_8,
+                verify_bitwise_xor_8_lookup_elements: interaction_elements
+                    .verify_bitwise_xor_8
+                    .clone(),
+            },
+            interaction_claim.verify_bitwise_xor_8.claimed_sum,
+        );
+        let verify_bitwise_xor_8_b_component = verify_bitwise_xor_8_b::Component::new(
+            tree_span_provider,
+            verify_bitwise_xor_8_b::Eval {
+                claim: cairo_claim.verify_bitwise_xor_8_b,
+                verify_bitwise_xor_8_b_lookup_elements: interaction_elements
+                    .verify_bitwise_xor_8_b
+                    .clone(),
+            },
+            interaction_claim.verify_bitwise_xor_8_b.claimed_sum,
+        );
+        let verify_bitwise_xor_9_component = verify_bitwise_xor_9::Component::new(
+            tree_span_provider,
+            verify_bitwise_xor_9::Eval {
+                claim: cairo_claim.verify_bitwise_xor_9,
+                verify_bitwise_xor_9_lookup_elements: interaction_elements
+                    .verify_bitwise_xor_9
+                    .clone(),
+            },
+            interaction_claim.verify_bitwise_xor_9.claimed_sum,
+        );
+        Self {
+            opcodes: opcode_components,
+            verify_instruction: verify_instruction_component,
+            blake_context,
+            builtins: builtin_components,
+            pedersen_context,
+            poseidon_context,
+            memory_address_to_id: memory_address_to_id_component,
+            memory_id_to_value: (
+                memory_id_to_value_components,
+                small_memory_id_to_value_component,
+            ),
+            range_checks: range_checks_component,
+            verify_bitwise_xor_4: verify_bitwise_xor_4_component,
+            verify_bitwise_xor_7: verify_bitwise_xor_7_component,
+            verify_bitwise_xor_8: verify_bitwise_xor_8_component,
+            verify_bitwise_xor_8_b: verify_bitwise_xor_8_b_component,
+            verify_bitwise_xor_9: verify_bitwise_xor_9_component,
+        }
+    }
+
+    pub fn provers(&self) -> Vec<&dyn ComponentProver<SimdBackend>> {
+        chain!(
+            self.opcodes.provers(),
+            [&self.verify_instruction as &dyn ComponentProver<SimdBackend>,],
+            self.blake_context.provers(),
+            self.builtins.provers(),
+            self.pedersen_context.provers(),
+            self.poseidon_context.provers(),
+            [&self.memory_address_to_id as &dyn ComponentProver<SimdBackend>,],
+            self.memory_id_to_value
+                .0
+                .iter()
+                .map(|component| component as &dyn ComponentProver<SimdBackend>),
+            [&self.memory_id_to_value.1 as &dyn ComponentProver<SimdBackend>,],
+            self.range_checks.provers(),
+            [
+                &self.verify_bitwise_xor_4 as &dyn ComponentProver<SimdBackend>,
+                &self.verify_bitwise_xor_7 as &dyn ComponentProver<SimdBackend>,
+                &self.verify_bitwise_xor_8 as &dyn ComponentProver<SimdBackend>,
+                &self.verify_bitwise_xor_8_b as &dyn ComponentProver<SimdBackend>,
+                &self.verify_bitwise_xor_9 as &dyn ComponentProver<SimdBackend>,
+            ]
+        )
+        .collect()
+    }
+
+    pub fn components(&self) -> Vec<&dyn Component> {
+        self.provers()
+            .into_iter()
+            .map(|component| component as &dyn Component)
+            .collect()
+    }
+}
+
+impl std::fmt::Display for CairoComponents {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        writeln!(f, "CairoComponents")?;
+        writeln!(f, "Opcodes: {}", self.opcodes)?;
+        writeln!(
+            f,
+            "VerifyInstruction: {}",
+            indented_component_display(&self.verify_instruction)
+        )?;
+        writeln!(f, "BlakeContext: {}", self.blake_context)?;
+        writeln!(f, "Builtins: {}", self.builtins)?;
+        writeln!(f, "PedersenContext: {}", self.pedersen_context)?;
+        writeln!(f, "PoseidonContext: {}", self.poseidon_context)?;
+        writeln!(
+            f,
+            "MemoryAddressToId: {}",
+            indented_component_display(&self.memory_address_to_id)
+        )?;
+        for component in &self.memory_id_to_value.0 {
+            writeln!(
+                f,
+                "MemoryIdToValue: {}",
+                indented_component_display(component)
+            )?;
+        }
+        writeln!(
+            f,
+            "SmallMemoryIdToValue: {}",
+            indented_component_display(&self.memory_id_to_value.1)
+        )?;
+        writeln!(f, "RangeChecks: {}", self.range_checks)?;
+        writeln!(
+            f,
+            "VerifyBitwiseXor4: {}",
+            indented_component_display(&self.verify_bitwise_xor_4)
+        )?;
+        writeln!(
+            f,
+            "VerifyBitwiseXor7: {}",
+            indented_component_display(&self.verify_bitwise_xor_7)
+        )?;
+        writeln!(
+            f,
+            "VerifyBitwiseXor8: {}",
+            indented_component_display(&self.verify_bitwise_xor_8)
+        )?;
+        writeln!(
+            f,
+            "VerifyBitwiseXor8B: {}",
+            indented_component_display(&self.verify_bitwise_xor_8_b)
+        )?;
+        writeln!(
+            f,
+            "VerifyBitwiseXor9: {}",
+            indented_component_display(&self.verify_bitwise_xor_9)
+        )?;
+        Ok(())
+    }
+}
