+++ stwo_cairo_prover/crates/cairo-air/src/lib.rs
@@ -2,9 +2,14 @@
 #![feature(array_chunks)]
 use serde::{Deserialize, Serialize};
 
-pub mod air;
 pub mod blake;
 pub mod builtins_air;
+pub mod cairo_proof;
+pub mod cairo_components;
+pub mod cairo_claim;
+pub mod cairo_interaction_elements;
+pub mod public_data;
+pub mod cairo_interaction_claim;
 pub mod components;
 pub mod opcodes_air;
 pub mod pedersen;
@@ -16,7 +21,7 @@ pub mod utils;
 // TODO(Ohad): verifier crate.
 pub mod verifier;
 
-pub use air::CairoProof;
+pub use cairo_proof::CairoProof;
 use stwo_cairo_common::preprocessed_columns::preprocessed_trace::PreProcessedTrace;
 
 /// The preprocessed trace used for the prover.
