+++ stwo_cairo_prover/crates/cairo-air/src/opcodes_air.rs
@@ -9,9 +9,9 @@ use stwo::prover::ComponentProver;
 use stwo_cairo_serialize::{CairoDeserialize, CairoSerialize};
 use stwo_constraint_framework::TraceLocationAllocator;
 
-use super::air::CairoInteractionElements;
+use crate::cairo_interaction_elements::CairoInteractionElements;
 use super::components::display_components;
-use crate::air::{accumulate_relation_uses, RelationUsesDict};
+use crate::cairo_claim::{accumulate_relation_uses, RelationUsesDict};
 use crate::components::{
     add_ap_opcode, add_opcode, add_opcode_small, assert_eq_opcode, assert_eq_opcode_double_deref,
     assert_eq_opcode_imm, blake_compress_opcode, call_opcode_abs, call_opcode_rel_imm,
