+++ stwo_cairo_prover/crates/prover/src/witness/utils.rs
@@ -1,6 +1,6 @@
 use std::sync::atomic::{AtomicU32, Ordering};
 
-use cairo_air::air::CairoClaim;
+use cairo_air::cairo_claim::CairoClaim;
 use cairo_air::PreProcessedTraceVariant;
 use itertools::Itertools;
 use num_traits::{One, Zero};
